AWS-IAM-Solution/requirements.txt
aws-cdk-lib>=2.0.0
constructs>=10.0.0
boto3>=1.26.0
python-jose>=3.3.0
aws-cdk.aws-sso>=2.0.0
